requirements.txt
aiohttp==3.9.1
feedparser==6.0.11
beautifulsoup4==4.12.3
python-dateutil==2.9.0
lxml==5.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0